Continue work on shell
